refactor(ui/*) : 입력 기능, 출력 기능 리팩토링
입력 기능, 출력 기능의 폴더 구조를 변경했습니다.

- ui 패키지에서 입력 검증 관리
- ui 패키지에서 출력 기능 관리
