+++ src/main/java/ui/InputView.java
@@ -1,4 +1,4 @@
-package bridge;
+package ui;
 
 /**
  * 사용자로부터 입력을 받는 역할을 한다.
